sources: refactor local source into module.
Extract the Local source into `sources/_local.py`.

LP: #1648906
